% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clyap.R
\name{clyap}
\alias{clyap}
\title{Solve continuous-time Lyapunov/Sylvester equations
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{AXE' + EXA'+Q=0.}}
\usage{
clyap(A, Q, E = NULL)
}
\arguments{
\item{A}{Square matrix}

\item{Q}{Square matrix}

\item{E}{Square matrix or \code{NULL}}
}
\value{
The solution matrix \eqn{X}.
}
\description{
Solve continuous-time Lyapunov/Sylvester equations
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{AXE' + EXA'+Q=0.}
}
\details{
`clyap` use `BKDIS` from ALGORITHM 705, COLLECTED ALGORITHMS FROM ACM
}
\examples{


}
