% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clyap.R
\name{clyap}
\alias{clyap}
\title{Solve continuous-time Lyapunov equations
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{BX+ XB'+Q=0.}}
\usage{
clyap(B, Q, E = NULL)
}
\arguments{
\item{B, }{Q: Conformable suqare matrices.}
}
\value{
The solution matrix \eqn{X}.
}
\description{
Solve continuous-time Lyapunov equations
\code{clyap} solve the continuous-time Lyapunov equations
\deqn{BX+ XB'+Q=0.}
}
\details{
`clyap` use `BKDIS` from ALGORITHM 705, COLLECTED ALGORITHMS FROM ACM
}
\examples{


}
